[Chore] #234 - 키보드 내리기 함수 위치 수정
